3.1-drf-intro/smart_home/requirements.txt
django~=5.0.7
djangorestframework~=3.15.2
psycopg2-binary

python-dotenv~=1.0.1